Update tests and remove extra files
